+++ spec/nio/selector_spec.rb
@@ -98,6 +98,28 @@ describe NIO::Selector do
       selected.should include(readable_monitor)
       selected.should_not include(unreadable_monitor)
     end
+
+    it "iterates across selected objects with a block" do
+      readable1, writer = IO.pipe
+      writer << "ohai"
+
+      readable2, writer = IO.pipe
+      writer << "ohai"
+
+      unreadable, _ = IO.pipe
+
+      monitor1 = subject.register(readable1, :r)
+      monitor2 = subject.register(readable2, :r)
+      monitor3 = subject.register(unreadable, :r)
+
+      readables = []
+      result = subject.select { |monitor| readables << monitor }
+      result.should == 2
+
+      readables.should include(monitor1)
+      readables.should include(monitor2)
+      readables.should_not include(monitor3)
+    end
   end
 
   context "select_each" do